Change type of 'ret' in one_read

+++ src/dev_one.c
@@ -20,7 +20,7 @@ static struct device *one_device = 0;
 static ssize_t one_read(struct file *file, char __user *buf, size_t len,
 			loff_t *offset)
 {
-	int ret = len;
+	ssize_t ret = len;
 
 	char *buff = kmalloc(len, GFP_KERNEL);
 	if (!buff)
